// frontend/src/api/generated/AdminOverview.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PoolStats } from "./PoolStats";
import type { QueueStats } from "./QueueStats";
import type { SystemCounts } from "./SystemCounts";

export type AdminOverview = { counts: SystemCounts, jobs: Array<QueueStats>, pool: PoolStats, };