// frontend/src/api/generated/AdminUserDetail.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { AdminUserInfo } from "./AdminUserInfo";
import type { MyOrg } from "./MyOrg";

export type AdminUserDetail = { user: AdminUserInfo, organizations: Array<MyOrg>, active_sessions: number, };