// frontend/src/api/generated/ApiKeyRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.

export type ApiKeyRow = { id: string, key_id: string, name: string, scopes: Array<string>, created_by: string | null, created_at: string, expires_at: string | null, last_used_at: string | null, revoked_at: string | null, };