// frontend/src/api/generated/OrgOverview.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { OrgRow } from "./OrgRow";
import type { OrgWidgets } from "./OrgWidgets";

export type OrgOverview = { organization: OrgRow, permissions: Array<string>, widgets: OrgWidgets, days: number, };