// frontend/src/api/generated/OrgRow.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { JsonValue } from "./serde_json/JsonValue";

export type OrgRow = { id: string, slug: string, name: string, personal: boolean, settings: JsonValue, billing_plan: string, billing_customer_ref: string | null, created_at: string, updated_at: string, deleted_at: string | null, };