// frontend/src/api/generated/RolesModelResponse.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { PermissionDescription } from "./PermissionDescription";
import type { RoleModel } from "./RoleModel";

export type RolesModelResponse = { organization_roles: Array<RoleModel>, system_roles: Array<RoleModel>, permissions: Array<PermissionDescription>, engine: string, system_roles_source: string, };